Cache raw CuPy kernels
Compiling a kernel on every call to _raw_apply_1d is costly - this
commit creates a cache of compiled raw kernels. The cache is initialized
on the first call to _raw_apply_1d, and and subsequent calls used the
cached kernels.

With this commit, I see the following timings for the resample_poly
example from the signaltools_examples notebook:

Numba:

 CPU times: user 288 ms, sys: 52 ms, total: 340 ms
 Wall time: 363 ms

CuPy:

 CPU times: user 116 ms, sys: 4 ms, total: 120 ms
 Wall time: 120 ms

A 3x speedup for the raw kernel compared to Numba.