Use float64 data types for resample_poly
Tests seem to fail with values that differ from
scipy.signal.resample_poly for both the Numba and raw kernels when using
a dtype of float32. This commit amends the CUDA kernel back to inferring
types (so when passed a float64 it will type the function for float64)
and the raw kernel to use doubles. This appears to fix all tests.

This commit also adds testing of the raw kernel as well as the default
Numba one.

Timings for a simple test:

    start = 0
    stop = 10
    num = int(1e8)
    resample_up = 2
    resample_down = 3

    gx = cp.linspace(start, stop, num, endpoint=False)
    gy = cp.cos(-gx**2/6.0)

    cusignal.resample_poly(gy, resample_up, resample_down,
                           window=('kaiser', 0.5),
                           use_numba=use_numba)

with this commit are:

use_numba=True::
    14.3 ms ± 2.03 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

use_numba=False:
    12.2 ms ± 25.5 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)