Preserve array datatype when zero-padding
Without specifying a datatype, the filter coefficients in resample_poly
were padded by concatenating arrays of float64. This caused the datatype
of the coefficients to change to whatever the numpy.result_type() of
the original type + float64 would be. This then determines the output
type of the filter, which caused the output type to change in ways
that appeared quite surprising to the user.